Backend/Model/requirements.txt
psycopg2-binary
pandas
google-generativeai
SQLAlchemy
python-dotenv
Flask
Flask-Cors
